+++ lib/hcloud/entities/certificate_status.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module HCloud
+  class CertificateStatus < Entity
+    attribute :error, :error
+    attribute :issuance
+    attribute :renewal
+  end
+end