+++ lib/hcloud/entities/used_by.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+module HCloud
+  class UsedBy < Entity
+    attribute :id, :integer
+    attribute :type
+  end
+end