+++ spec/support/integration.rb
@@ -34,6 +34,7 @@ RSpec.configure do |config|
     logger.level = :warn
 
     # Clean up cloud resources
+    HCloud::Certificate.all.each(&:delete)
     HCloud::Server.all.each(&:delete)
     HCloud::Network.all.each(&:delete)
     HCloud::PlacementGroup.all.each(&:delete)