+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Implemented Networks
 - Implemented Network Actions
 - Implemented Pricing
+- Implemented Certificates
 
 ## HCloud v1.1.0